Implement plugin system for EmailService.
This commit introduces a plugin system for the EmailService in the
arbeitszeitapp. The new system empowers administrators to customize
email delivery methods according to their requirements. Notably, this
enhancement allows for independent testing of email functionality,
facilitating the eventual removal of the dependency on `flask_mail`
from our codebase.
